+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.2] - 2024-12-17
+**No Breaking Changes. Safe to Bump**
+### DirectoryStructure Class
+- Added `get_files()` method to return a list of DirectoryItems that are all of metadata type `file`
+- Added `get_directories()` method to return a list of DirectoryItems that are all of metadata type `directory`
+
+### Summarization
+- Added more detailed `INFO` logs including directory size and files being summarized (optional argument)
+
+
 ## [0.2.1] - 2024-12-17
 **No Breaking Changes. Safe to Bump**
 ### Directory Parser
@@ -7,7 +17,7 @@
 - Renamed `parse_from_directory_structure` to `parse_from_style`. Old method still valid until **v0.3.0**.
 - Added `template_to_directory_structure` method to convert templates to DirectoryStructure objects
 
-### DirectoryItem
+### DirectoryItem Class
 - Changed order in which how Metadata appears in dict
 
 ### Style Changes